requirements.txt
Django==2.2.11
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
